requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.7
django-allauth==0.63.6
django-bootstrap-icons==0.9.0
django-chunked-upload==2.0.0
django-import-export==4.1.1
djangorestframework==3.15.2
idna==3.7
kombu==5.4.2
numpy==2.1.1
pandas==2.2.3
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websockets==13.1
